+++ ChangeLog.md
@@ -15,6 +15,7 @@
     - Fix DataTables dt-container display style to be visible below dropdown menus
     - Private Table._setCheckboxes() method is renamed _setTemplatesFromStrings() and is more extensible
     - Add datatables.net-plugins NPM dependency
+    - Replace assertions par logger.error() + exception throw
 
 ### 1.6.0